+++ src/database/operations/opportunities/index.ts
@@ -1,3 +1,4 @@
 export * from './createOpportunity';
 export * from './loadOpportunity';
 export * from './loadOpportunityBundle';
+export * from './loadSystemOpportunity';