+++ src/database/queries/applicationForms/selectSystemApplicationForm.sql
@@ -1,0 +1,3 @@
+SELECT application_form_to_json(application_forms.*) AS object
+FROM application_forms
+WHERE id = system_application_form_id();