+++ src/database/queries/bulkUploadTasks/selectById.sql
@@ -5,11 +5,15 @@ SELECT
 		:authContextIsAdministrator
 	) AS object
 FROM bulk_upload_tasks
+	INNER JOIN application_forms
+		ON bulk_upload_tasks.application_form_id = application_forms.id
+	INNER JOIN opportunities
+		ON application_forms.opportunity_id = opportunities.id
 WHERE
-	id = :bulkUploadTaskId
+	bulk_upload_tasks.id = :bulkUploadTaskId
 	AND has_funder_permission(
 		:authContextKeycloakUserId,
 		:authContextIsAdministrator,
-		funder_short_code,
+		opportunities.funder_short_code,
 		'view'
 	);